get user profile implemented
