+++ src/constants/types.ts
@@ -1,8 +1,20 @@
+import { Types } from "mongoose";
+
 type LoginRequestBody = {
     email: string;
     password: string;
 }
 
+type UserProfileResponse = {
+    _id: Types.ObjectId;
+    firstName: string;
+    lastName?: string;
+    email: string;
+    password: string;
+    role: "user" | "admin";
+}
+
 export {
-    LoginRequestBody
+    LoginRequestBody,
+    UserProfileResponse
 }
