+++ src/controller/userCtrl.ts
@@ -1,51 +1,29 @@
 import { Request, Response } from 'express';
 import User from '../models/userModels';
 import expressAsyncHandler from 'express-async-handler';
-import type { LoginRequestBody } from 'constants/types';
-import { generateToken } from '../config/jwtToken';
 
-// Create a new user
-const createUser = expressAsyncHandler(
+// Get user profile
+const getProfile = expressAsyncHandler(
     async (req: Request, res: Response) => {
-        const { email } = req.body;
-        const findUser = await User.findOne({ email });
-        if (!findUser) {
-            // Create a new user
-            const newUser = await User.create(req.body);
-            res.json(newUser);
-        }
-        else {
-            // User already exist
-            throw new Error("User already exists!");
+        const { id: userId } = req.query;
+        if (!req.query.id) throw new Error("Plz provide userId");
+        try {
+            const findUser = await User.findById(userId).select("-password");
+            if (findUser) {
+                // Send User Profile without Except Password
+                res.send(findUser);
+            }
+            else {
+                // User already exist
+                throw new Error("User doesn't exist");
+            }
+        } catch (error: any) {
+            throw new Error(error);
         }
     }
 );
 
-// Login existing user
-const loginUser = expressAsyncHandler(
-    async (req: Request, res: Response) => {
-        const { email, password }: LoginRequestBody = req.body;
-        // console.log({ email, password });
 
-        // Check if user exist or not
-        const findUser = await User.findOne({ email });
-        if (findUser) {
-            if (await findUser.isPasswordMatched(password)) {
-                const findUserRes = findUser.toJSON(); // converting MongoDB Object to a normal Object
-                const { password, ...rest } = findUserRes; // removing "password" field from response
-                res.json({
-                    ...rest,
-                    token: generateToken(findUser._id)
-                }); // sending user detail without "password"
-            }
-            else {
-                throw new Error("Incorrect Password!");
-            }
-        }
-        else {
-            throw new Error("User doesn't exist!");
-        }
-    }
-)
 
-export { createUser, loginUser };
+
+export { getProfile };
